# auth/src/static/openapi.yml
openapi: 3.0.0
info:
  title: Auth service
  description: API для авторизации и создания ролей
  version: 1.0.0
tags:
  - name: auth
    description: API для сайта и личного кабинета
  - name: roles
    description: API для управления доступами
  - name: user roles
    description: API для ролей у пользователя
paths:
  /api/v1/account/create_user:
    post:
      tags:
        - auth
      summary: Создать нового пользователя
      description: Создать нового пользователя
      operationId: create_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: Successful operation
        '400':
          description: You have to pass login and password
        '409':
          description: User with this login already exists
  /api/v1/account/login:
    post:
      tags:
        - auth
      summary: Войти в аккаунт
      description: Войти в аккаунт
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: You have to pass login and password
        '401':
          description: Wrong login or password
  /api/v1/account/refresh:
    post:
      tags:
        - auth
      summary: Обновить токен
      description: Обновить токен
      operationId: refresh
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessToken'
        '401':
          description: Missing Authorization Header
        '422':
          description: Wrong token
  /api/v1/account/logout:
    delete:
      tags:
        - auth
      summary: Выйти из аккаунта
      description: Выйти из аккаунта
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token has been revoked
        '401':
          description: Missing Authorization Header
        '422':
          description: Wrong token
  /api/v1/account/change_login:
    put:
      tags:
        - auth
      summary: Изменить логин
      description: Изменить логин
      operationId: change_login
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
      responses:
        '200':
          description: Successful operation
        '401':
          description: Missing Authorization Header
        '409':
          description: User with this login already exists
        '422':
          description: Wrong token
  /api/v1/account/change_password:
    put:
      tags:
        - auth
      summary: Изменить пароль
      description: Изменить пароль
      operationId: change_password
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Password'
      responses:
        '200':
          description: Successful operation
        '401':
          description: Missing Authorization Header
        '422':
          description: Wrong token
  /api/v1/account/history:
    get:
      tags:
        - auth
      summary: История входов
      description: История входов
      operationId: history
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/History'
        '401':
          description: Missing Authorization Header
        '422':
          description: Wrong token
  /api/v1/roles/create:
    post:
      tags:
        - roles
      summary: Создать роль
      description: Создать роль
      operationId: create_role
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseRole'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          description: Токен невалидный
        '400':
          description: Роль с таким именем уже существует
        '403':
          description: Отказано в доступе
  /api/v1/roles/delete/{role_id}:
    delete:
      tags:
        - roles
      summary: Удалить роль
      description: Удалить роль
      operationId: delete_role
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: role_id
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '200':
          description: Successful operation
        '401':
          description: Токен невалидный
        '400':
          description: Роль с таким id не существует
        '403':
          description: Отказано в доступе
  /api/v1/roles/change/{role_id}:
    put:
      tags:
        - roles
      summary: Изменить роль
      description: Изменить роль
      operationId: change_role
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: role_id
          schema:
            type: string
            format: uuid
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseRole'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          description: Токен невалидный
        '400':
          description: Роль с таким именем уже существует
        '403':
          description: Отказано в доступе
  /api/v1/roles:
    get:
      tags:
        - roles
      summary: Посмотреть все роли
      description: Посмотреть все роли
      operationId: roles
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Roles'
        '401':
          description: Токен невалидный
        '403':
          description: Отказано в доступе
  /api/v1/roles/{user_id}:
    get:
      tags:
        - user roles
      summary: Посмотреть все роли пользователя
      description: Посмотреть все роли пользователя
      operationId: user_roles
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Roles'
        '401':
          description: Токен невалидный
        '403':
          description: Отказано в доступе
  /api/v1/roles/{user_id}/create:
    post:
      tags:
        - user roles
      summary: Назначить роль пользователю
      description: Назначить роль пользователю
      operationId: user_create_role
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
            format: uuid
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseRole'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          description: Токен невалидный
        '400':
          description: Роль с таким именем уже существует
        '403':
          description: Отказано в доступе
  /api/v1/roles/{user_id}/delete:
    delete:
      tags:
        - user roles
      summary: Удалить у пользователя роль
      description: Удалить у пользователя роль
      operationId: user_delete_role
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
            format: uuid
          required: true
        - in: query
          name: role_name
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful operation
        '401':
          description: Токен невалидный
        '400':
          description: Роль с таким id не существует
        '403':
          description: Отказано в доступе
components:
  schemas:
    User:
      type: object
      required:
        - login
        - password
      properties:
        login:
          type: string
          example: best_user
        password:
          type: string
          example: best_password
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: 123
        refresh_token:
          type: string
          example: 123
    AccessToken:
      type: object
      properties:
        access_token:
          type: string
          example: 123
    RefreshToken:
      type: object
      properties:
        refresh_token:
          type: string
          example: 123
    Login:
      type: object
      properties:
        login:
          type: string
          example: 'best_login'
    Password:
      type: object
      properties:
        password:
          type: string
          example: 'best_password'
    Log:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user:
          type: string
          example: 'best_user'
        timestamp:
          type: integer
          example: 1675847177
    History:
      type: array
      items:
        $ref: '#/components/schemas/Log'
    BaseRole:
      type: object
      properties:
        name:
          type: string
          example: user
    Role:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: user
    Roles:
      type: array
      items:
        $ref: '#/components/schemas/Role'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []